backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlmodel>=0.0.21
yfinance>=0.2.54
apscheduler>=3.10
requests>=2.32
beautifulsoup4>=4.12
python-dotenv>=1.0
google-genai>=1.0
